positionInColumn changed to FLOAT8/Double

+++ src/main/java/com/stodo/projectchaos/model/entity/TaskEntity.java
@@ -21,7 +21,7 @@ public class TaskEntity extends Auditable {
     private String title;
 
     private String description;
-    private Short positionInColumn;
+    private Double positionInColumn;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "assignee_email")
